+++ src/Glpi/Search/Output/Spreadsheet.php
@@ -104,16 +104,6 @@ abstract class Spreadsheet extends ExportSearchOutput
             return false;
         }
 
-        // Clear output buffers
-        $ob_config = ini_get('output_buffering');
-        $max_level = filter_var($ob_config, FILTER_VALIDATE_BOOLEAN) ? 1 : 0;
-        while (ob_get_level() > $max_level) {
-            ob_end_clean();
-        }
-        if (ob_get_level() > 0) {
-            ob_clean();
-        }
-
         $spread = $this->getSpreasheet();
         $writer = $this->getWriter();